fix bugs in subtitles and unzip
